pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.3.0"
description = "Сервис, который анализирует указанные страницы на SEO-пригодность"
authors = ["Anna Bystrova <annbystr@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Aluwian/python-project-83"
packages = [{include = "page_analyzer"}]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Topic :: Internet :: WWW/HTTP :: Site Management :: Link Checking"
]


[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
validators = "^0.20.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.3.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
